refactor: use job-based prediction panel
